Fix various size_t related warnings and errors
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@26035 a1c6a512-1295-4272-9138-f99709370657

+++ apps/buffering.c
@@ -685,8 +685,8 @@ static bool buffer_handle(int handle_id)
             /* stop buffering data for now and post-pone buffering the rest */
             stop = true;
             DEBUGF( "%s(): Preventing handle corruption: h1.id:%d h2.id:%d"
-                    " copy_n:%lu overlap:%ld h1.filerem:%lu\n", __func__,
-                    h->id, h->next->id, (unsigned long)copy_n, (long)overlap,
+                    " copy_n:%zu overlap:%ld h1.filerem:%lu\n", __func__,
+                    h->id, h->next->id, (unsigned long)copy_n, overlap,
                     (unsigned long)h->filerem);
             copy_n -= overlap;
         }
